pfch_Stock Ticker Capital Appreciation Comparison
This project focused on creating a platform for visualizations for information on impact investment. The Objective of this projects was to use various coding languages and platforms to visualize data in a more digestible way by none tech savvy audiences. It is important to display financial information in an understandable manner. Stocks Tickers performance are a method of getting a clear view of a companies performance in real time. For this project apart from the HTML and CSS used in bootstrap to build the webpages, Python and Tableau were the tools used. This Visualization was made to give a visual representation of Adjusted close performances of Stocks within the year 2018 from months July to November. For this visualization the Stocks that were retrieved were AAPL - Apple, MSFT -Microsoft, CSCO -Cisco, CRM - Salesforce, BBY- Best buy.

python - python5.py -uses python to call data via na API the uses Matplotlib to prints the visualization for the pages. 
               python33.py - prints out Ad.Close csv to import into tableau.
csv - import into tableau for further visualizations.

html - htm files, create pages for the website